+++ tests/integration/test_cli.py
@@ -23,9 +23,7 @@ from dffml.service.dev import Develop
 from dffml.util.packaging import is_develop
 from dffml.util.entrypoint import load
 from dffml.config.config import BaseConfigLoader
-from dffml.util.asynctestcase import AsyncTestCase
-
-from .common import IntegrationCLITestCase
+from dffml.util.asynctestcase import AsyncTestCase, IntegrationCLITestCase
 
 
 class TestList(IntegrationCLITestCase):